+++ src/xpcs/filter/sparse_filter.cpp
@@ -104,13 +104,13 @@ SparseFilter::~SparseFilter() {
 
 void SparseFilter::Apply(struct xpcs::io::ImmBlock* blk) {
   int **indx = blk->index;
-  short **val = blk->value;
+  float **val = blk->value;
   int frames = blk->frames;
   std::vector<int> ppf = blk->pixels_per_frame;
   for (int i = 0; i < frames; i++) {
     int pixels = ppf[i];
     int *index = indx[i];
-    short *value = val[i];
+    float *value = val[i];
     float f_sum = 0.0f;
     int pix_cnt = frame_width_ * frame_height_;
     int sbin = 0;
@@ -118,6 +118,7 @@ void SparseFilter::Apply(struct xpcs::io::ImmBlock* blk) {
     if (frame_index_ > 0 && (frame_index_ % static_window_) == 0) {
       partition_no_++;
     }
+
     for (int j = 0; j < pixels; j++) {
 
       if (pixel_mask_[index[j]] != 0) {
